+++ libs/components/avatar/src/app/public/public_api.ts
@@ -1,3 +1,0 @@
-export * from './modules/avatar/avatar-size';
-export * from './modules/avatar/avatar-src';
-export * from './modules/avatar/avatar.module';