+++ libs/components/avatar/CHANGELOG.md
@@ -1,6 +1,24 @@
+# 5.0.0-beta.3 (2021-9-16)
+
+- Updated peer dependencies. [#64](https://github.com/blackbaud/skyux-avatar/pull/64)
+
+# 5.0.0-beta.2 (2021-09-09)
+
+- Updated peer dependencies. [#61](https://github.com/blackbaud/skyux-avatar/pull/61)
+
+# 5.0.0-beta.1 (2021-09-02)
+
+- Migrated to Angular CLI. [#59](https://github.com/blackbaud/skyux-avatar/pull/59)
+
 # 4.1.4 (2021-07-13)
 
 - Fixed the avatar component to let users load files with a parenthesis in the filename. [#57](https://github.com/blackbaud/skyux-avatar/pull/57)
+
+# 5.0.0-beta.0 (2021-07-08)
+
+- Initial beta release.
+- Added support for `@angular/core@^12`. [#56](https://github.com/blackbaud/skyux-avatar/pull/56)
+- Added support for `5.0.0-beta.*` versions of SKY UX component libraries. [#56](https://github.com/blackbaud/skyux-avatar/pull/56)
 
 # 4.1.3 (2021-03-17)